+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - OpenGraph meta tags for Suite (web version)
 - Option to hide dashboard graph
 - Instructions for verifying linux binary (landing page)
+- Ethereum replace-by-fee feature
 
 ### Changed
 - Don't allow to run multiple instances of Suite desktop app